Versiyon2/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jsonschema_pydantic==0.6
keyboard==0.13.5
langchain==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain-text-splitters==0.3.9
langsmith==0.4.9
Markdown==3.8.2
mcp==1.12.3
mcp-use==1.3.7
multidict==6.6.3
orjson==3.11.1
packaging==25.0
pillow==10.4.0
posthog==6.3.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
scarf-sdk==0.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
tkhtmlview==0.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
